Adding download option
Maybe you have a reason to not adding this, but if you ran "dataset = MNISTDataModule('.')" in a fresh install, the reported error is "RuntimeError: Dataset not found. You can use download=True to download it", which is confusing (because it is needed to run "dataset.prepare_dataset()". Thus either add this option or change the error report. Adding this to the tutorials docs may help also.