Warnings fix (#449)
* Revert "Merge pull request #1 from ganprad/warnings_fix"

This reverts commit 7c5aaf04148fb4e5b8128219479ba75c1364e9aa.

* Fixes warning related np.integer in SklearnDataModule

Fixes this warning:
```DeprecationWarning: Converting `np.integer` or `np.signedinteger` to a dtype is deprecated. The current result is `np.dtype(np.int_)` which is not strictly correct. Note that the result depends on the system. To ensure stable results use may want to use `np.int64` or `np.int32````